BUILD : 1.7.5 (423) BUGS : N/A NOTES : code tidy up fixes stuff under Freebsd
git-svn-id: svn://svn.anope.org/anope/trunk@423 31f1291d-b8d6-0310-a050-a5561fc1590b


git-svn-id: http://anope.svn.sourceforge.net/svnroot/anope/trunk@281 5417fbe8-f217-4b02-8779-1006273d7864

+++ src/bahamut.c
@@ -733,7 +733,8 @@ void anope_cmd_part(char *nick, char *chan, const char *fmt, ...)
 void anope_cmd_topic(char *whosets, char *chan, char *whosetit,
                      char *topic, time_t when)
 {
-    send_cmd(whosets, "TOPIC %s %s %lu :%s", chan, whosetit, when, topic);
+    send_cmd(whosets, "TOPIC %s %s %lu :%s", chan, whosetit,
+             (unsigned long int) when, topic);
 }
 
 /* UNSQLINE */
@@ -745,7 +746,7 @@ void anope_cmd_unsqline(char *user)
 /* JOIN - SJOIN */
 void anope_cmd_join(char *user, char *channel, time_t chantime)
 {
-    send_cmd(user, "SJOIN %ld %s", chantime, channel);
+    send_cmd(user, "SJOIN %ld %s", (long int) chantime, channel);
 }
 
 /* AKILL */
@@ -760,7 +761,7 @@ void anope_cmd_akill(char *user, char *host, char *who, time_t when,
                      time_t expires, char *reason)
 {
     send_cmd(NULL, "AKILL %s %s %d %s %ld :%s", host, user, 86400 * 2, who,
-             time(NULL), reason);
+             (long int) time(NULL), reason);
 }
 
 /* SVSKILL */
@@ -802,8 +803,9 @@ void anope_cmd_svskill(char *source, char *user, const char *fmt, ...)
  */
 void anope_cmd_svsmode(User * u, int ac, char **av)
 {
-    send_cmd(ServerName, "SVSMODE %s %ld %s%s%s", u->nick, u->timestamp,
-             av[0], (ac == 2 ? " " : ""), (ac == 2 ? av[1] : ""));
+    send_cmd(ServerName, "SVSMODE %s %ld %s%s%s", u->nick,
+             (long int) u->timestamp, av[0], (ac == 2 ? " " : ""),
+             (ac == 2 ? av[1] : ""));
 }
 
 /* SQUIT */
@@ -849,7 +851,7 @@ void anope_cmd_connect(int servernum)
  */
 void anope_cmd_svinfo()
 {
-    send_cmd(NULL, "SVINFO 3 1 0 :%ld", time(NULL));
+    send_cmd(NULL, "SVINFO 3 1 0 :%ld", (long int) time(NULL));
 }
 
 /* PASS */
@@ -1296,8 +1298,9 @@ void anope_cmd_211(const char *fmt, ...)
 void anope_cmd_nick(char *nick, char *name, char *modes)
 {
     EnforceQlinedNick(nick, NULL);
-    send_cmd(NULL, "NICK %s 1 %ld %s %s %s %s 0 0 :%s", nick, time(NULL),
-             modes, ServiceUser, ServiceHost, ServerName, name);
+    send_cmd(NULL, "NICK %s 1 %ld %s %s %s %s 0 0 :%s", nick,
+             (long int) time(NULL), modes, ServiceUser, ServiceHost,
+             ServerName, name);
     anope_cmd_sqline(nick, "Reserved for services");
 }
 
@@ -1404,8 +1407,8 @@ void anope_cmd_bot_nick(char *nick, char *user, char *host, char *real,
                         char *modes)
 {
     EnforceQlinedNick(nick, s_BotServ);
-    send_cmd(NULL, "NICK %s 1 %ld %s %s %s %s 0 0 :%s", nick, time(NULL),
-             modes, user, host, ServerName, real);
+    send_cmd(NULL, "NICK %s 1 %ld %s %s %s %s 0 0 :%s", nick,
+             (long int) time(NULL), modes, user, host, ServerName, real);
     anope_cmd_sqline(nick, "Reserved for services");
 }
 
@@ -1420,14 +1423,14 @@ void anope_cmd_svsnick(char *source, char *guest, time_t when)
     if (!source || !guest) {
         return;
     }
-    send_cmd(NULL, "SVSNICK %s %s :%ld", source, guest, when);
+    send_cmd(NULL, "SVSNICK %s %s :%ld", source, guest, (long int) when);
 }
 
 void anope_cmd_guest_nick(char *nick, char *user, char *host, char *real,
                           char *modes)
 {
-    send_cmd(NULL, "NICK %s 1 %ld %s %s %s %s 0 0 :%s", nick, time(NULL),
-             modes, user, host, ServerName, real);
+    send_cmd(NULL, "NICK %s 1 %ld %s %s %s %s 0 0 :%s", nick,
+             (long int) time(NULL), modes, user, host, ServerName, real);
 }
 
 void anope_cmd_svso(char *source, char *nick, char *flag)
@@ -1449,7 +1452,8 @@ void anope_cmd_vhost_off(User * u)
 /* sent if svid is something weird */
 void anope_cmd_svid_umode(char *nick, time_t ts)
 {
-    send_cmd(ServerName, "SVSMODE %s %lu +d 1", nick, ts);
+    send_cmd(ServerName, "SVSMODE %s %lu +d 1", nick,
+             (unsigned long int) ts);
 }
