+++ src/ultimate2.c
@@ -781,7 +781,8 @@ void anope_cmd_remove_akill(char *user, char *host)
 void anope_cmd_topic(char *whosets, char *chan, char *whosetit,
                      char *topic, time_t when)
 {
-    send_cmd(whosets, "TOPIC %s %s %lu :%s", chan, whosetit, when, topic);
+    send_cmd(whosets, "TOPIC %s %s %lu :%s", chan, whosetit,
+             (unsigned long int) when, topic);
 }
 
 void anope_cmd_vhost_off(User * u)
@@ -1180,7 +1181,7 @@ void anope_cmd_svsnick(char *source, char *guest, time_t when)
     if (!source || !guest) {
         return;
     }
-    send_cmd(NULL, "SVSNICK %s %s :%ld", source, guest, when);
+    send_cmd(NULL, "SVSNICK %s %s :%ld", source, guest, (long int) when);
 }
 
 /* Events */
@@ -1418,8 +1419,9 @@ void anope_cmd_serv_privmsg(char *source, char *dest, char *msg)
 void anope_cmd_nick(char *nick, char *name, char *mode)
 {
     EnforceQlinedNick(nick, NULL);
-    send_cmd(NULL, "NICK %s 1 %ld %s %s %s 0 :%s", nick, time(NULL),
-             ServiceUser, ServiceHost, ServerName, name);
+    send_cmd(NULL, "NICK %s 1 %ld %s %s %s 0 :%s", nick,
+             (long int) time(NULL), ServiceUser, ServiceHost, ServerName,
+             name);
     anope_cmd_mode(nick, nick, "%s", mode);
     anope_cmd_sqline(nick, "Reserved for services");
 }
@@ -1473,8 +1475,8 @@ void anope_cmd_bot_nick(char *nick, char *user, char *host, char *real,
                         char *modes)
 {
     EnforceQlinedNick(nick, s_BotServ);
-    send_cmd(NULL, "NICK %s 1 %ld %s %s %s 0 :%s", nick, time(NULL),
-             user, host, ServerName, real);
+    send_cmd(NULL, "NICK %s 1 %ld %s %s %s 0 :%s", nick,
+             (long int) time(NULL), user, host, ServerName, real);
     anope_cmd_mode(nick, nick, "%s", modes);
     anope_cmd_sqline(nick, "Reserved for services");
 }
@@ -1545,8 +1547,8 @@ void anope_cmd_sgline(char *mask, char *reason)
 void anope_cmd_guest_nick(char *nick, char *user, char *host, char *real,
                           char *modes)
 {
-    send_cmd(NULL, "NICK %s 1 %ld %s %s %s 0 :%s", nick, time(NULL),
-             user, host, ServerName, real);
+    send_cmd(NULL, "NICK %s 1 %ld %s %s %s 0 :%s", nick,
+             (long int) time(NULL), user, host, ServerName, real);
     anope_cmd_mode(nick, "MODE %s %s", nick, modes);
 }
